Process Scheduling Simulation
Testing and comparing different scheduling algorithms to be used in different scenarios. It helps in choosing an appropriate algorithm both for an OS or
large scale system. It aids in the selection of an appropriate algorithm for an operating system or large-scale systems. On such systems, user/process 
access patterns to resources (e.g., CPU, memory) and system metrics (e.g., process turnaround time, CPU/memory usage) can be recorded for later analysis.
Data from a real system can also be used as an input to simulate real-world processes/applications and test how the system would perform with different 
scheduling algorithms. 

In this repository, I've impleted three different scheduling algorithms, each with it's unique functionality.  

Each of these program will be taking a filepath as its only arguement, and should read be reading process data from the